Work around JAX ConcretizationTypeError
JAX doesn't like it when the shape of some value depends on a
traced variable. Because f is traced (even though its size will never
change), a ConcretizationTypeError was being thrown when calculating
in_shape. This is easily worked around by more explicitly calculating
the values of m and n